*More code cleanup.  Although I don't usually bother with the third party source files, I removed unused variables from some of the AGUP ones because they now get warnings in the latest GCC, which would have annoyed me to no end. *Updated GH import so that pitch values of 2 are treated as pitch shifts. *Updated INI export so that the pro drums not present tag is written if no pro drum difficulty is defined and there are no cymbals, and a vocal harmonies not present tag if no vocal harmony difficulty is defined and there are no harmony tracks stored into the project.

+++ src/editor.c
@@ -55,11 +55,11 @@ float eof_pos_distance(float p1, float p2)
 	}
 }
 
-void eof_select_beat(int beat)
+void eof_select_beat(unsigned long beat)
 {
 	eof_log("eof_select_beat() entered", 1);
 
-	int i;
+	unsigned long i;
 	int beat_counter = 0;
 	char first_measure = 0;
 	eof_selected_beat = beat;
